rename files in validation folder
